Starting add support for 'conversations'
Rather than relying on a seperate mechanism to send messages to
multiple clients - or 'Group Chats'. I've instead opted for a
'conversation' mechanismm that allows for multiple participants.
In effect, creating the the same functionality.

In this commit, user A creates a new conversation with user
A & B, or A & B & C.... The amount of pariticipants is variable. This
could be altered in the future to cap the amount of participants
in a conversation as well as the amount of conversations the
user can be a part of.