WilsonServer: Updated interfaces names
I prefer the standard of prefacing interfaces with 'I', i.e.,
Cat interface would be 'ICat' however, in TypeScript interfaces
are used to define extensible types. It doesn't make sense for
a consumer of the API to import an interface to define a type.

Instead, I opt to only use the prefacing 'I' when creating an
interface that has the intention of actually being implmented by
somthing i.e., 'IWilsonClientConversationManager' etc... Not
'IConversation' for example.