WilsonServer: Implemented conversation invites
To prevent users maliciously changing who participates in a 'conversation', I implemented an invite only mechanism that clients can respond to.

If user A wants to talk to user A or B or both, they send an invite event to the server with those user(s) as the participants.

The participants are then notified and can reject or accept to request.

This conversation is then updated in the client's session as accepted or rejected and the connected client can then receive messages from that conversation.